+++ cove/src/euph/small_message.rs
@@ -1,15 +1,20 @@
+use std::hash::{DefaultHasher, Hash, Hasher};
+
 use crossterm::style::Stylize;
-use euphoxide::api::{MessageId, Snowflake, Time};
+use euphoxide::api::{MessageId, Snowflake, Time, UserId};
 use jiff::Timestamp;
 use toss::{Style, Styled};
 
 use crate::{store::Msg, ui::ChatMsg};
+
+use super::util;
 
 #[derive(Debug, Clone)]
 pub struct SmallMessage {
     pub id: MessageId,
     pub parent: Option<MessageId>,
     pub time: Time,
+    pub user_id: UserId,
     pub nick: String,
     pub content: String,
     pub seen: bool,
@@ -69,6 +74,14 @@ impl Msg for SmallMessage {
 
     fn last_possible_id() -> Self::Id {
         MessageId(Snowflake::MAX)
+    }
+
+    fn nick_emoji(&self) -> Option<String> {
+        let mut hasher = DefaultHasher::new();
+        self.user_id.0.hash(&mut hasher);
+        let hash = hasher.finish();
+        let emoji = &util::EMOJI_LIST[hash as usize % util::EMOJI_LIST.len()];
+        Some(emoji.clone())
     }
 }